requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
redis==5.0.1
aiosmtplib==2.0.0
dnspython==2.4.2
pydantic-settings==2.0.3
aiohttp==3.9.1
python-dotenv==1.0.0
email-validator==2.0.0